+++ src/main/java/com/sparta/hotitemcollector/domain/like/LikeService.java
@@ -1,18 +1,15 @@
 package com.sparta.hotitemcollector.domain.like;
 
 import com.sparta.hotitemcollector.domain.product.Product;
-import com.sparta.hotitemcollector.domain.product.ProductRepository;
 import com.sparta.hotitemcollector.domain.product.ProductService;
+import com.sparta.hotitemcollector.domain.product.ProductSimpleResponseDto;
 import com.sparta.hotitemcollector.domain.user.User;
-import com.sparta.hotitemcollector.global.exception.CustomException;
-import com.sparta.hotitemcollector.global.exception.ErrorCode;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -28,6 +25,7 @@ public class LikeService {
         if (existingLike.isPresent()) {
             // 좋아요가 이미 존재하면 삭제
             likeRepository.delete(existingLike.get());
+            productService.decreaseLikes(productId);
             return false; // 좋아요 취소
         } else {
             // 좋아요가 존재하지 않으면 생성
@@ -39,8 +37,18 @@ public class LikeService {
                     .build();
 
             likeRepository.save(like);
+            productService.increaseLikes(productId);
             return true; // 좋아요 생성
         }
+    }
+
+    @Transactional(readOnly = true)
+    public List<ProductSimpleResponseDto> getLikeProduct(User user) {
+        List<Product> productList = findLikeProductIdByUser(user);
+
+        return productList.stream()
+            .map(ProductSimpleResponseDto::new)
+            .collect(Collectors.toList());
     }
 
     @Transactional(readOnly = true)
